SVGs kiiiind of work
Using the library svg-android, I was able to display SVGs as an
imagebutton. Some SVGs do, such as the loading image. Some don't, such
as the play and pause images. This is wierd, and I can't figure out why
they get different treatment. However, it should be noted that play and
pause images are about 100x bigger in size than the loading image.